% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{bs}
\alias{bs}
\title{Backward substitution with backsolve}
\usage{
bs(Up)
}
\arguments{
\item{Up}{an upper triangular matrix}
}
\description{
bs: the inverse of a upper triangular matrix with backsolve
}
\examples{
Up <- matrix(c(2, 0, 0, 3, 4, 0, 3, 2, 5), nrow = 3, ncol = 3)
bs(Up) #
solve(Up)
}
