% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fs}
\alias{fs}
\title{forward substitution with forwardsolve}
\usage{
fs(Lw)
}
\arguments{
\item{Lw}{a lower triangular matrix}
}
\description{
fs: the inverse of a lower triangular matrix with forwardsolve
}
\examples{
Lw <- matrix(c(3, 2, 1, 0, 4, 1, 0, 0, 2), nrow = 3, ncol = 3)
fs(lw) #
solve(Lw)
}
