% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_recursion.R
\name{kalman_filter_recursion}
\alias{kalman_filter_recursion}
\title{kalman_filter_recursion}
\usage{
kalman_filter_recursion(y, u, x0, P0, F, E, H, V, W)
}
\arguments{
\item{y}{The observation vector (\eqn{l \times 1})}

\item{u}{The exogenous vector (\eqn{n \times 1})}

\item{x0}{The initial state vector (\eqn{k \times 1})}

\item{P0}{The initial state covariance matrix (\eqn{k \times k})}

\item{F}{The state transition matrix (\eqn{k \times k})}

\item{E}{The control matrix (\eqn{k \times n})}

\item{H}{The observation matrix (\eqn{l \times k})}

\item{V}{The state noise covariance matrix (\eqn{k \times k})}

\item{W}{The observation noise covariance matrix (\eqn{l \times l})}
}
\value{
A list with the following components:
\item{xt}{The estimated state vector (\eqn{k \times 1})}
\item{Pt}{The estimated state covariance matrix (\eqn{k \times k})}
}
\description{
kalman_filter_recursion is the main function to implement
the Kalman filter.
}
\details{
The Kalman filter is a recursive algorithm for estimating
the state of a linear dynamic system from a series of observed
measurements.

The state space model is given by
\deqn{x(t) = F x(t-1) + E u(t) + v(t)}
\deqn{y(t) = H x(t) + w(t)}
where \eqn{x(t)} is the state vector (\eqn{k \times 1}),
\eqn{y(t)} is the observation vector (\eqn{l \times 1}),
\eqn{u(t)} is the exogenous vector (\eqn{n \times 1}),
\eqn{v(t)} is the state noise (\eqn{k \times 1}),
and \eqn{w(t)} is the observation noise (\eqn{l \times 1}).
The Kalman filter is used to estimate the state vector
\eqn{x(t)} given the observations \eqn{y(t)}.

The Kalman filter is implemented using the following recursion:
Prediction step:
 \deqn{x(t|t-1) = F x(t-1|t-1) + E u(t)}
 \deqn{P(t|t-1) = F P(t-1|t-1) F^t + V}
 Innovation step:
 \deqn{e(t) = y(t) - H x(t|t-1)}
 \deqn{s(t) = H P(t|t-1) H^t + W}
 Update step:
 \deqn{K(t) = P(t|t-1) H^t s(t)^{-1}}
 \deqn{x(t|t) = x(t|t-1) + K(t) e(t)}
 \deqn{P(t|t) = (I - K(t) H) P(t|t-1)}
 where \eqn{x(t|t)} is the estimated state vector at time t given
 the observations up to time t,
 \eqn{P(t|t)} is the estimated state covariance matrix
 at time t given the observations up to time t,
 \eqn{K(t)} is the Kalman gain at time t,
See Kitagawa, (2010), Introduction to Time Series Modeling, Chapman & Hall
for more details.
}
