% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_kalman_filter.R
\name{qr_r}
\alias{qr_r}
\title{qr_kalman_filter}
\usage{
qr_r(AA, BB)
}
\description{
Square root Kalman Filter using only QR decompositions.
}
\details{
This package provides a square root Kalman filter implementation
that uses only QR decompositions. This is a numerically stable way to
implement the Kalman filter.
}
\author{
Koichi (Koiti) Yano
}
