% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_kalman_filter_recursion.R
\name{qr_kalman_filter_recursion}
\alias{qr_kalman_filter_recursion}
\title{qr_kalman_filter_recursion}
\usage{
qr_kalman_filter_recursion(y, u = 0, x0, Sig0, F, E, H, Gm_v, Gm_w)
}
\arguments{
\item{y}{The observation vector (\eqn{l \times 1})}

\item{u}{The exogenous vector (\eqn{n \times 1})}

\item{x0}{The initial state vector (\eqn{k \times 1})}

\item{Sig0}{The square root of the initial state covariance matrix (\eqn{k \times k})}

\item{F}{The state transition matrix (\eqn{k \times k})}

\item{E}{The control matrix (\eqn{k \times n})}

\item{H}{The observation matrix (\eqn{l \times k})}

\item{Gm_v}{The square root of the state noise covariance matrix (\eqn{k \times k})}

\item{Gm_w}{The square root of the observation noise covariance matrix (\eqn{l \times l})}
}
\value{
A list with the following components:
\item{x(t)}{The estimated state vector (\eqn{k \times 1})}
\item{Sig(t)}{The square root of the estimated state covariance matrix (\eqn{k \times k})}
}
\description{
qr_kalman_filter_recursion is the main function to implement
the square root Kalman Filter using only QR decompositions.
}
\details{
The state space model is given by
\deqn{x(t) = F x(t-1) + E u(t) + v(t)}
\deqn{y(t) = H x(t) + w(t)}
where \eqn{x(t)} is the state vector (\eqn{k \times 1}),
\eqn{y(t)} is the observation vector (\eqn{l \times 1}),
\eqn{u(t)} is the exogenous vector (\eqn{n \times 1}),
\eqn{v(t)} is the state noise (\eqn{k \times 1}),
and \eqn{w(t)} is the observation noise (\eqn{l \times 1}).
The square root Kalman Filter is used to estimate the state vector
\eqn{x(t)} given the observations \eqn{y(t)}.

The QR Kalman filter (square root Kalman Filter using only QR
decompositions) is implemented using the following recursion:
Prediction step:
 \deqn{x(t|t-1) = F x(t-1|t-1) + E u(t)}
 \deqn{\Sigma(t|t-1) = gr_r(\Sigma(t-1|t-1) F^t, \Gamma_v)}
 Innovation step:
 \deqn{e(t) = y(t) - H x(t|t-1)}
 \deqn{G(t) = gr_r(\Sigma (t|t-1) H^t, \Gamma_w)}
 Update step:
 \deqn{K(t) = {[G(t)^{-1} (G(t)^{-t}  H^t) { \Sigma (t|t-1)}^t \Sigma(t|t-1)]}^{t}}
 \deqn{x(t|t) = x(t|t-1) + K(t) e(t)}
 \deqn{\Sigma(t|t) = gr_r(\Sigma(t|t-1) {(I - K(t) H)}^t, \Gamma_w {K(t)}^t)}
 where \eqn{x(t|t)} is the estimated state vector,
 \eqn{G(t)} is the square root of the estimated state covariance matrix,
 and \eqn{K(t)} is the Kalman gain at time t.

 See Tracy, (2022), "A Square-Root Kalman Filter Using Only QR
 Decompositions," for more details.
}
